Add TypeScript config for server, fix MongoDB connection error
