+++ server/nodemon.json
@@ -1,0 +1,5 @@
+{
+    "watch": ["src"],
+    "ext": "ts",
+    "exec": "ts-node-esm src/server.ts"
+}
