+++ server/src/db/db.ts
@@ -1,13 +1,14 @@
-const mongoose = require('mongoose');
+import mongoose from "mongoose";
 
 const db = async () => {
   try {
     mongoose.set('strictQuery', false);
-    await mongoose.connect(process.env.MONGODB_URL);
+    await mongoose.connect(process.env.MONGODB_URL as string || '');
+  
     console.log('Connected to MongoDB');
   } catch (error) {
     console.error('Failed to connect to MongoDB', error);
   }
 };
 
-module.exports = db;
+export default db;
