+++ server/src/server.ts
@@ -1,0 +1,19 @@
+// Import required modules
+import http from 'http';
+import dotenv from 'dotenv';
+// Load environment variables (Jote: dotenv.config() must be top of the file to be activated correctly)
+dotenv.config();
+
+import app from './app';
+
+
+
+// Define the port
+const PORT = process.env.PORT || 5001; // if there is no port, return port 5000
+
+// create server instance
+const server = http.createServer(app);
+
+server.listen(PORT, () => {
+    console.log(`Server started on port ${PORT}.`)
+});
